// src/app/client-view/models/current-scenario/retirement-income-goal.model.ts
export class RetirementIncomeGoalsModel {
    goalId: string;
    name: string;
    retirementYear: number;
    retirementAge: number;
    target: number;
    currentIncome: number;
    currentCapital: number;
    desireRetirementIncome: number;
    description: string;
    adviserComment: string;
    category: number;
    subCategory: number;
    goalType: number;
    type: number;
    onTrack: number;
}